requirements.txt
bcrypt
fasteners
git+https://github.com/ryanvilbrandt/bottle.git@master
bottle_websocket
google-api-python-client